fix:pet-detail-ui
