+++ src/app/(products)/_components/image.tsx
@@ -1,0 +1,9 @@
+import Image from "next/image";
+
+export default function ProductImage() {
+  return (
+    <div className="relative w-full pb-[80%]">
+      <Image src="/assets/images/food-detail.jpg" fill alt="whole hearted" />
+    </div>
+  );
+}
