+++ src/app/(products)/_shared/tab.ts
@@ -1,0 +1,7 @@
+export const Tabs = {
+  DESCRIPTION: 1,
+  ADDITIONAL_INFO: 2,
+  REVIEWS: 3,
+};
+
+export type TabsType = (typeof Tabs)[keyof typeof Tabs];
