+++ src/app/(products)/foods/_components/page-content.tsx
