+++ src/app/(products)/pets/_components/color-checkbox.tsx
