+++ src/app/(products)/pets/_components/ingredient-checkbox.tsx
