+++ src/app/(products)/pets/_components/pets-category.tsx
