+++ src/app/(products)/supplies/_components/page-content.tsx
