+++ src/app/(products)/supplies/_components/supplies-category.tsx
