+++ src/app/(products)/supplies/page.tsx
