+++ src/components/supply-card.tsx
@@ -15,7 +15,7 @@ export default function SupplyCard() {
   return (
     <div className="border-box flex min-w-[232px] flex-1 transform flex-col border border-solid border-light_gray_color_second bg-white small-screen:min-w-[25%] x-small-screen:min-w-[calc(100%/3)] x-smallest-screen:min-w-[50%]">
       <div className="relative w-full overflow-hidden pb-[85%]">
-        <Link href="#" className="h-full w-full">
+        <Link href="/supplies/564845321" className="h-full w-full">
           <Image
             src="/assets/images/food1.jpg"
             alt="food1"
